Edit

+++ WishList/Controllers/ItemController.cs
@@ -40,7 +40,10 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
-            var item = _context.Items.First(i => i.Id == id);
+            var item = _context.Items.FirstOrDefault(i => i.Id == id);
+            if (item == null)
+                return View("Error");
+
             _context.Items.Remove(item);
             _context.SaveChanges();
